Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/url-resolution-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: The given initial value does not parse for the given syntax.
CONSOLE MESSAGE: line 2: SyntaxError: The given initial value does not parse for the given syntax.
CONSOLE MESSAGE: line 1515: Error: Tried to create a test with file_is_test

FAIL url-resolution SyntaxError: The given initial value does not parse for the given syntax.
